backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
confection==0.1.5
contextualSpellCheck==0.4.4
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
editdistance==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.0
fsspec==2024.6.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
grpcio
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.0.0
ipykernel==6.29.4
ipython==8.25.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
namex==0.0.8
nest_asyncio==1.6.0
networkx==3.3
nltk==3.8.2
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.5.5.64
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
regex==2024.7.24
requests==2.32.2
requests-oauthlib==1.3.0
rich==13.7.1
rsa==4.7.2
safetensors==0.4.4
scikit-learn==1.4.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
sounddevice==0.4.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
sympy==1.13.2
tensorboard==2.16.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow_decision_forests==1.9.0
termcolor==2.1.0
textblob==0.18.0.post0
textdistance==4.6.3
tf_keras==2.16.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
uvicorn==0.20.0
wasabi==1.1.3
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.35.1
wrapt==1.14.1
wurlitzer==3.1.1
yarl==1.9.3
zipp==3.19.2
